requirements.txt
streamlit>=1.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.0.6
pandas>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
python-decouple>=3.8
pytest>=8.0.0
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0 